fix(Questions routes): split prod and serv routes, one handler that identify path
