+++ PF_Api/src/routes/questServRouter.js
@@ -1,9 +1,10 @@
 const { Router } = require("express");
-const { getQuestionsHandler, postQuestionsHandler } = require("../handlers/questProdHandlers.js");
+const { getQuestionsHandler, postQuestionsHandler } = require("../handlers/questionsHandlers.js");
 const questServRouter = Router();
 
-questServRouter.get("/", getQuestionsHandler );
-questServRouter.post("/:offerId", postQuestionsHandler );
+const prod_path = '/service'
+questServRouter.get(prod_path, getQuestionsHandler );
+questServRouter.post(`${prod_path}/:offerId`, postQuestionsHandler );
 
 
 module.exports = questServRouter;
